testing: Add recv-bench.hs
On my system, the results are approximately:

 * Winsock.recv: 39us

 * Network.Socket.ByteString.recv: 2.4us

Thus, our recv takes 15 times as long as the current implementation.

It seems that sending the work to another thread, then waiting for the result
in yet another thread, is quite expensive.  I don't mind too badly, but it
might be a regression for some people.